refactor(search): move catalog index types out of the SQLite module
The catalog projection's document, snapshot, result, and hit types, the
query-term normalization, the searchable-text composition, and the
closed vocabularies for surface kinds and field roles move from
`catalog::sqlite_index` to a backend-neutral `catalog::index`. The
retrieval class now names document kinds instead of carrying SQL, so a
second storage backend can share the vocabulary and render its own
predicate.

No behavior change; the SQLite index keeps its statements and tests.
