+++ crates/coral-app/src/search/catalog/mod.rs
@@ -1,5 +1,6 @@
 //! Catalog metadata search provider.
 
+pub(crate) mod index;
 pub(crate) mod provider;
 pub(crate) mod snapshot;
 pub(crate) mod sqlite_index;
